Use SaveableListener instead of onConfigurationChange
