+++ src/test/groovy/hudson/plugins/gradle/injection/ExpectedException.java
@@ -1,0 +1,8 @@
+package hudson.plugins.gradle.injection;
+
+public class ExpectedException extends RuntimeException {
+
+    public ExpectedException() {
+        super("expected");
+    }
+}
